[PAM-3456] Do not use regexp for URL validation
Universal URL matching is somewhat complex syntactically, so RegExp becomes messy and unreadable.
In addition, the current version has cases where it hangs the browser due to computational
complexity for certain inputs.

Instead, just parse the URL using browser built-in routines and check that protocol is either http
or https, to validate a link.